requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0